re-organized files and made variant more variatic
