+++ singularity-opac/base/opac_error.hpp
